requirements.txt
python-dotenv
playwright
openai
faiss-cpu
tiktoken
pyyaml